update: rename AutoLocalizationManager to AutoTranslationManager and add Auto Translation Tool
