--- Editor/Localization/LocalizationContextMenu.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 09bcbf4fde8299545b66c6c0c96bdef7
