--- Editor/Localization/README.md.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 34eb69b720c687242a798e3250a2372a
-TextScriptImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
