--- Editor/Localization/SimpleLocalizationDashboard.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: d7e20d25c855434459d02f0efdd440a0
