--- Scripts/Localization/RuntimeLocalizationHelper.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 63da6d18fe33f6c4fa2fe6785fd5ddba
